% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_convert_images.R
\name{convert_images}
\alias{convert_images}
\title{Convert image format.}
\usage{
convert_images(image, format = "png")
}
\arguments{
\item{image}{The image to be converted. \code{getPerlin} writes images in
.bmp format which are not always what's required.}

\item{format}{The format argument to \code{magick::image_write}. Usually
"png".}
}
\value{
Nothing. Converts between image formats.
}
\description{
Convert image format.
}
