% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getPerlin}
\alias{getPerlin}
\title{Prints Perlin noise to file.}
\usage{
getPerlin(nOctaves, frequency, landSize, isWrapped, outputPath)
}
\arguments{
\item{nOctaves}{Number of octaves. May be thought of as small scale 
variability. Must be an integer value between 1 and 8. Higher values
result in landscapes with more small scale noise.}

\item{frequency}{Frequency of noise. May be thought of as large scale
variability. May be any double value between 1.0 and 16.0. Higher values
result in noise that is closer to white noise.}

\item{landSize}{Landscape size in pixels. The image saved will be a
square with this side.}

\item{isWrapped}{A boolean value indicating whether the landscape is
tiled, that is, whether the pattern repeats. Useful when generating
landscapes for simulations.}

\item{outputPath}{A string specifying the image file to write the landscape.}
}
\value{
Nothing. Writes an image to file.
}
\description{
Prints Perlin noise to file.
}
