% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pln_get_noise}
\alias{pln_get_noise}
\title{Prints Perlin noise to file.}
\usage{
pln_get_noise(
  n_octaves,
  frequency,
  increment,
  save_image,
  image_pixels,
  wrapped,
  output_path
)
}
\arguments{
\item{n_octaves}{Number of octaves. May be thought of as small scale 
variability. Must be an integer value between 1 and 8. Higher values
result in landscapes with more small scale noise.}

\item{frequency}{Frequency of noise. May be thought of as large scale
variability. May be any double value between 1.0 and 16.0. Higher values
result in noise that is closer to white noise.}

\item{increment}{The increment at which to move through the landscape.}

\item{save_image}{A boolean of whether to save the image.}

\item{image_pixels}{Landscape size in pixels. The image saved will be a
square with this side.}

\item{wrapped}{A boolean value indicating whether the landscape is
tiled, that is, whether the pattern repeats. Useful when generating
landscapes for simulations.}

\item{output_path}{The image filepath to write the landscape.}
}
\value{
A matrix of Perlin noise. Optionally, writes an image to file.
}
\description{
Prints Perlin noise to file.
}
