% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proclandr-package.R
\docType{package}
\name{proclandr-package}
\alias{proclandr}
\alias{proclandr-package}
\title{proclandr: Wrapped Perlin Noise Landscapes in R}
\description{
Wrapped Perlin noise landscapes in R using libnoise.
}
\author{
\strong{Maintainer}: Pratik Gupte \email{p.r.gupte@rug.nl} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID})

}
\keyword{internal}
