% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proclandr-package.R
\docType{package}
\name{proclandr-package}
\alias{proclandr}
\alias{proclandr-package}
\title{proclandr: Wrapped Perlin Noise Landscapes in R}
\description{
Wrapped Perlin noise landscapes in R using libnoise.
}
\keyword{internal}
